requirements.txt
tensorflow
numpy
matplotlib
flask
pillow
pyttsx3
speechrecognition
pyaudio
tk
